requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alibabacloud-credentials==0.3.6
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.3
alibabacloud-imagerecog20190930==2.0.10
alibabacloud-imagesearch20210501==1.2.3
alibabacloud-openapi-util==0.2.2
alibabacloud-openplatform20191219==2.0.0
alibabacloud-oss-sdk==0.1.0
alibabacloud-oss-util==0.0.6
alibabacloud-tea==0.4.3
alibabacloud-tea-fileform==0.0.5
alibabacloud-tea-openapi==0.3.13
alibabacloud-tea-util==0.3.13
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.16.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jmespath==0.10.0
multidict==6.4.3
openai==1.73.0
packaging==25.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
realtime==2.4.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
websockets==14.2
yarl==1.19.0
